==> 02_embedding-sentences.ipynb <==
## Job Crafting: Embedding Sentences

from sentence_transformers import SentenceTransformer

import numpy as np
import pandas as pd

DATA = 'data/'

# load sentence transformer model
# https://huggingface.co/sentence-transformers/distiluse-base-multilingual-cased-v2
model = SentenceTransformer('distiluse-base-multilingual-cased-v2')

# load raw strategy sentences
data = pd.read_csv(DATA + 'strategies_raw.csv', sep=';')
strategies = data['strategy text'].tolist()

# embed strategy sentences
strategy_embeddings = model.encode(strategies, show_progress_bar=True)

# save strategy embeddings to file
# np.save(DATA + 'strategy_embeddings', strategy_embeddings)  # use previous embeddings for full replicability

# read vignette texts from file
with open(DATA + 'vignettes.txt') as f:
    vignettes = f.read()
    
vignettes = vignettes.split('; ')

# embed vignettes
vignette_embeddings = model.encode(vignettes)

# save vignette embeddings to file
# np.save(DATA + 'vignette_embeddings', vignette_embeddings)  # use previous embeddings for full replicability
