==> 04_reducing-item-bias.ipynb <==
## Job Crafting: Reducing Item Bias

from sentence_transformers import SentenceTransformer

import numpy as np
import pandas as pd
import re

DATA = 'data/'

data = pd.read_csv(DATA + 'strategies_raw.csv', sep=';')

model = SentenceTransformer('distiluse-base-multilingual-cased-v2')

masks =  {1: ['tiempo', 'trabaj*', 'situacion*'], 
          2: ['cambi*', 'puesto', 'nuev*', 'empresa', 'rend*', 'compañer*'],
          3: ['efica*', 'productiv*', 'trabaj*'],
          4: ['presión', 'estrés']}

data_masked = pd.DataFrame(columns=['strategy text', 'item'])

for i in range(len(data)):
    strategy_new = data.iloc[i]['strategy text']
    for mask in masks[data.iloc[i]['item']]:
        strategy_new = re.sub(mask, '[MASK]', strategy_new)
    data_masked.loc[i] = [strategy_new, data.iloc[i]['item']]

strategy_embeddings_masked = model.encode(data_masked['strategy text'].tolist(), show_progress_bar=True)

# np.save(DATA + 'strategy_embeddings_masked', strategy_embeddings_masked)
